# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.MenuItemCreate:
    properties:
      display_name:
        maxLength: 100
        minLength: 3
        type: string
      icon:
        maxLength: 50
        type: string
      is_active:
        type: boolean
      name:
        maxLength: 50
        minLength: 3
        type: string
      order_index:
        minimum: 0
        type: integer
      parent_id:
        type: string
      path:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - display_name
    - name
    - path
    type: object
  dto.MenuItemUpdate:
    properties:
      display_name:
        maxLength: 100
        minLength: 3
        type: string
      icon:
        maxLength: 50
        type: string
      is_active:
        type: boolean
      order_index:
        minimum: 0
        type: integer
      parent_id:
        type: string
      path:
        maxLength: 255
        minLength: 1
        type: string
    type: object
...
  response.Error:
    properties:
      code:
        type: integer
      data: {}
      error: {}
      log_id:
        type: string
      message:
        type: string
      status:
        type: boolean
    type: object
  response.Success:
    properties:
      code:
        type: integer
      data: {}
      error: {}
      log_id:
        type: string
      message:
        type: string
      status:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: zaiduszhuhur@gmail.com
    name: API Support
    url: https://www.linkedin.com/in/zaidus-zhuhur/
  description: This is a sample server for a safety riding service.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Safety Riding API
  version: "1.0"
paths:
...
  /menu:
    post:
      consumes:
      - application/json
      description: Create a new menu entry
      parameters:
      - description: Menu payload
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/dto.MenuItemCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - ApiKeyAuth: []
      summary: Create a menu
      tags:
      - Menus
  /menu/{id}:
    delete:
      consumes:
      - application/json
      description: Delete menu by ID
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - ApiKeyAuth: []
      summary: Delete a menu
      tags:
      - Menus
    get:
      consumes:
      - application/json
      description: Retrieve menu information by ID
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - ApiKeyAuth: []
      summary: Get menu detail
      tags:
      - Menus
    put:
      consumes:
      - application/json
      description: Update menu information by ID
      parameters:
      - description: Menu ID
        in: path
        name: id
        required: true
        type: string
      - description: Menu payload
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/dto.MenuItemUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - ApiKeyAuth: []
      summary: Update a menu
      tags:
      - Menus
  /menus:
    get:
      consumes:
      - application/json
      description: Retrieve paginated menus (admin only)
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - ApiKeyAuth: []
      summary: List menus
      tags:
      - Menus
  /menus/active:
    get:
      consumes:
      - application/json
      description: Retrieve menus that are marked active for current user role
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - ApiKeyAuth: []
      summary: Get active menus
      tags:
      - Menus
  /menus/me:
    get:
      consumes:
      - application/json
      description: Retrieve menus based on authenticated user's role
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Success'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Error'
      security:
      - ApiKeyAuth: []
      summary: Get menus for current user
      tags:
      - Menus
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"